requirements.txt
absl-py==1.4.0
asgiref==3.6.0
backports.zoneinfo==0.2.1
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
django-crispy-forms==2.0
django-icons==22.1
ffmpeg-python==0.2.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.51.3
idna
importlib-metadata==6.0.0
importlib-resources==5.12.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
numpy==1.24.2
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==4.22.0
psutil==5.9.4
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser 
pyOpenSSL
pyparsing==3.0.9
PySocks 
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
smmap==5.0.0
sqlparse==0.4.3
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
tzdata==2022.7
Werkzeug==2.2.3
wincertstore==0.2
zipp==3.15.0
